Move get_non_fungible_ids method into nonfungible blueprint
