+++ radix-engine/src/blueprints/resource/fungible/fungible_vault.rs
@@ -7,7 +7,7 @@ use crate::types::*;
 use native_sdk::runtime::Runtime;
 use radix_engine_interface::api::substate_api::LockFlags;
 use radix_engine_interface::api::ClientApi;
-use radix_engine_interface::api::{types::*, ClientSubstateApi};
+use radix_engine_interface::api::types::*;
 use radix_engine_interface::blueprints::resource::*;
 
 pub struct FungibleVaultBlueprint;
